refactor(map translation): move some logic out of MapTabContent to hooks
