+++ src/components/MapTranslatorTabs/MapTabContent.tsx
@@ -1,32 +1,27 @@
-import { IonItem, IonLabel, IonList, useIonAlert } from '@ionic/react';
+import { IonItem, IonLabel, IonList } from '@ionic/react';
 import { MouseEvent, useCallback, useEffect, useState } from 'react';
-import { type INode, parseSync } from 'svgson';
 import {
   MuiMaterial,
   Button,
   LanguageInfo,
   LangSelector,
 } from '@eten-lab/ui-kit';
-import { nanoid } from 'nanoid';
 import {
   StyledFilterButton,
   StyledSectionTypography,
 } from './StyledComponents';
 import { useAppContext } from '@/src/hooks/useAppContext';
-import { useMapTranslationTools } from '@/src/hooks/useMapTranslationTools';
+import {
+  MapDetail,
+  eProcessStatus,
+  useMapTranslationTools,
+} from '@/src/hooks/useMapTranslationTools';
 import { langInfo2String } from '@/utils/langUtils';
 import { FeedbackTypes } from '@/constants/common.constant';
 const { Box, styled, CircularProgress } = MuiMaterial;
 
 const PADDING = 20;
 
-enum eProcessStatus {
-  NONE = 'NONE',
-  PARSING_STARTED = 'PARSING_STARTED',
-  PARSING_COMPLETED = 'PARSING_COMPLETED',
-  COMPLETED = 'SAVED_IN_DB',
-  FAILED = 'FAILED',
-}
 enum eUploadMapBtnStatus {
   NONE,
   LANG_SELECTION,
@@ -34,15 +29,6 @@ enum eUploadMapBtnStatus {
   SAVING_FILE,
   COMPLETED,
 }
-type MapDetail = {
-  id?: string;
-  tempId?: string;
-  status: eProcessStatus;
-  words?: string[];
-  mapFileId?: string;
-  name?: string;
-  langInfo: LanguageInfo;
-};
 
 export const MapTabContent = () => {
   const {
@@ -55,10 +41,9 @@ export const MapTabContent = () => {
 
   const [mapList, setMapList] = useState<MapDetail[]>([]);
   const [langInfo, setLangInfo] = useState<LanguageInfo | undefined>();
-  const [presentAlert] = useIonAlert();
   const [uploadMapBtnStatus, setUploadMapBtnStatus] =
     useState<eUploadMapBtnStatus>(eUploadMapBtnStatus.NONE);
-  const { sendMapFile } = useMapTranslationTools();
+  const { processFile, setMapStatus } = useMapTranslationTools();
 
   // for now we show all maps despite selected language
   useEffect(() => {
@@ -82,159 +67,52 @@ export const MapTabContent = () => {
   }, [singletons]);
 
   useEffect(() => {
+    if (!singletons) return;
     for (const mapState of mapList) {
-      if (mapState.status === eProcessStatus.PARSING_COMPLETED) {
-        const processMapWords = async (
-          words: string[],
-          langInfo: LanguageInfo,
-          mapId: string,
-        ) => {
-          if (!singletons || !words.length || !langInfo) return;
-          let hasNextBatch = true;
-          let batchNumber = 0;
-          const batchItemCount = 100;
-          const createdWords = [];
-          while (hasNextBatch) {
-            const startIdx = batchNumber * batchItemCount;
-            const endIdx = startIdx + batchItemCount;
-            const batchWords = words.slice(startIdx, endIdx);
-            logger.info(
-              'hasNextBatch',
-              hasNextBatch,
-              startIdx,
-              endIdx,
-              batchWords,
-            );
-            createdWords.push(
-              ...(await singletons.wordService.createWordsWithLangForMap(
-                batchWords.map((w) => w.trim()).filter((w) => w !== ''),
-                langInfo,
-                mapId,
-              )),
-            );
-            if (batchWords.length !== batchItemCount) {
-              hasNextBatch = false;
-            }
-            batchNumber++;
-          }
-          logger.info('total created words', createdWords);
+      if (mapState.status !== eProcessStatus.PARSING_COMPLETED) continue;
+
+      const handleMapParsingCompleted = async (argMap: MapDetail) => {
+        const newState: Partial<MapDetail> = {
+          status: eProcessStatus.COMPLETED,
         };
-        const handleMapParsingCompleted = async (argMap: MapDetail) => {
-          if (!singletons) return;
-          const newState: Partial<MapDetail> = {
-            status: eProcessStatus.COMPLETED,
-          };
-          try {
-            const mapId = await singletons.mapService.saveMap(argMap.langInfo, {
-              name: argMap.name!,
-              mapFileId: argMap.mapFileId!,
-              ext: 'svg',
-            });
-            if (mapId) {
-              newState.id = mapId;
-              processMapWords(argMap.words!, argMap.langInfo, mapId);
-            } else newState.status = eProcessStatus.FAILED;
-          } catch (error) {
-            newState.status = eProcessStatus.FAILED;
-          }
-          setMapStatus(argMap.tempId!, newState);
-        };
-        handleMapParsingCompleted(mapState);
-      }
+        try {
+          const mapId = await singletons.mapService.saveMap(argMap.langInfo, {
+            name: argMap.name!,
+            mapFileId: argMap.mapFileId!,
+            ext: 'svg',
+          });
+          if (mapId) {
+            newState.id = mapId;
+            singletons.mapService.processMapWords(
+              argMap.words!,
+              argMap.langInfo,
+              mapId,
+            );
+          } else newState.status = eProcessStatus.FAILED;
+        } catch (error) {
+          newState.status = eProcessStatus.FAILED;
+        }
+        setMapStatus(argMap.tempId!, newState, setMapList);
+      };
+      handleMapParsingCompleted(mapState);
     }
-  }, [logger, mapList, singletons]);
-
-  const showAlert = useCallback(
-    (msg: string) => {
-      presentAlert({
-        header: 'Alert',
-        subHeader: 'Important Message!',
-        message: msg,
-        buttons: ['Ok'],
-      });
-    },
-    [presentAlert],
-  );
-
-  const setMapStatus = (tempId: string, state: Partial<MapDetail>) => {
-    setMapList((prevList) => {
-      const clonedList = [...prevList];
-      const idx = clonedList.findIndex((m) => m.tempId === tempId);
-      if (idx > -1) {
-        clonedList[idx] = { ...clonedList[idx], ...state };
-      }
-      return clonedList;
-    });
-  };
+  }, [logger, mapList, setMapStatus, singletons]);
 
   const fileHandler: React.ChangeEventHandler<HTMLInputElement> = useCallback(
     (e) => {
       const file = e.target.files?.[0];
-      if (file == null) return;
-      if (!langInfo) return;
-      const fileName = file.name?.split('.')[0];
-      const id = nanoid();
-      let alreadyExists = false;
-
-      setMapList((prevList) => {
-        const existingIdx = prevList.findIndex((map) => map.name === fileName);
-        if (existingIdx >= 0) {
-          alertFeedback(FeedbackTypes.ERROR, 'File already exists');
-          alreadyExists = true;
-          return [...prevList];
-        }
-        return [
-          ...prevList,
-          {
-            tempId: id,
-            name: fileName,
-            status: eProcessStatus.PARSING_STARTED,
-            langInfo,
-          },
-        ];
-      });
-      if (alreadyExists) return;
-      const fileReader = new FileReader();
-      fileReader.onload = function (evt: ProgressEvent<FileReader>) {
-        if (evt.target?.readyState !== 2) return;
-        if (evt.target.error != null) {
-          setMapStatus(id, { status: eProcessStatus.FAILED });
-          showAlert('Error while reading file. Read console for more info');
-          return;
-        }
-        const filecontent = evt.target.result;
-        if (!filecontent) {
-          setMapStatus(id, { status: eProcessStatus.FAILED });
-          showAlert('Error while reading file. Read console for more info');
-          return;
-        }
-        const originalSvg = filecontent.toString();
-        const parsed = parseSync(originalSvg);
-        const textArray: string[] = [];
-        iterateOverINode(parsed, ['style'], (node) => {
-          if (node.type === 'text' || node.type === 'textPath') {
-            if (!node.value) return;
-            textArray.push(node.value);
-          }
-        });
-
-        if (textArray.length === 0 && originalSvg) {
-          setMapStatus(id, { status: eProcessStatus.FAILED });
-          showAlert('No text or textPath tags found');
-        } else {
-          sendMapFile(file, (sentFileData) => {
-            setMapStatus(id, {
-              status: eProcessStatus.PARSING_COMPLETED,
-              mapFileId: sentFileData.id,
-              words: textArray,
-            });
-          });
-        }
-      };
-      fileReader.readAsText(file);
+      if (!file) {
+        alertFeedback(FeedbackTypes.ERROR, 'No file selected');
+        return;
+      }
+      if (!langInfo) {
+        alertFeedback(FeedbackTypes.ERROR, 'No language seleced');
+        return;
+      }
+      processFile(file, langInfo, setMapList);
       e.target.value = '';
     },
-    [alertFeedback, langInfo, sendMapFile, showAlert],
+    [alertFeedback, langInfo, processFile],
   );
 
   // for now we show all maps despite selected language
@@ -268,6 +146,7 @@ export const MapTabContent = () => {
       if (langInfo) {
         setUploadMapBtnStatus(eUploadMapBtnStatus.UPLOAD_FILE);
       } else {
+        alertFeedback(FeedbackTypes.INFO, 'Please select language first.');
         e.stopPropagation();
         e.preventDefault();
       }
@@ -402,27 +281,6 @@ export const MapTabContent = () => {
     </Box>
   );
 };
-
-/**
- * Should iterate over INode and its children in a consistent order
- * @param node starting node
- * @param skipNodeNames node names to exclude with all its children
- * @param cb callback to be applied
- * @returns
- */
-function iterateOverINode(
-  node: INode,
-  skipNodeNames: string[],
-  cb: (node: INode) => void,
-) {
-  if (skipNodeNames.includes(node.name)) return;
-
-  cb(node);
-
-  for (const child of node.children || []) {
-    iterateOverINode(child, skipNodeNames, cb);
-  }
-}
 
 const StyledBox = styled(Box)(() => ({
   width: '100%',
